requirements.txt
streamlit
torch
torchvision
Pillow
transformers
numpy